refactor(openfin): move OpenFinLogo under components
