--- src/client/src/rt-platforms/openfin-platform/adapter/logo.tsx
@@ -1,23 +1,0 @@
-import React from 'react'
-import { OpenFinLogoLink } from './styled'
-import { OpenFinLogo } from './assets/OpenFinLogo'
-import { OpenFinBrowserLink } from '../components'
-import styled from 'styled-components/macro'
-
-export default () => {
-  const Logo = styled.div`
-    margin-right: 0.75rem;
-  `
-
-  return (
-    <Logo>
-      <OpenFinLogoLink
-        href="http://www.openfin.co"
-        as={OpenFinBrowserLink}
-        data-qa="logo__openfin-logo-link"
-      >
-        <OpenFinLogo />
-      </OpenFinLogoLink>
-    </Logo>
-  )
-}
