--- src/client/src/rt-platforms/openfin-platform/adapter/styled.tsx
@@ -1,7 +1,0 @@
-import styled from 'styled-components/macro'
-
-export const OpenFinLogoLink = styled.a`
-  .svg-fill {
-    fill: ${({ theme }) => theme.core.textColor};
-  }
-`
