+++ src/client/src/rt-platforms/openfin-platform/components/index.ts
@@ -5,4 +5,5 @@ export {
   OpenFinHeader,
   OPENFIN_CHROME_HEADER_HEIGHT,
 } from './OpenFinChrome'
+export { OpenFinLogo } from './OpenFinLogo'
 export { default as OpenFinBrowserLink } from './OpenFinBrowserLink'
